Made a file utility class: Moved functions from program.cs to class FileUtility.cs Write to file Read file Check file
