+++ BasicSecurity Crypto Program/BasicSecurity Crypto Program/BasicSecurity Crypto Program.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Cryptos\SecurityAes.cs" />
     <Compile Include="User\User.cs" />
+    <Compile Include="Utility\FileUtility.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
